+++ internal/oci/ecr/credentials_store_test.go
@@ -1,0 +1,93 @@
+package ecr
+
+import (
+	"context"
+	"encoding/base64"
+	"io"
+	"testing"
+	"time"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
+	"github.com/stretchr/testify/require"
+	"oras.land/oras-go/v2/registry/remote/auth"
+)
+
+func TestDefaultClientFunc(t *testing.T) {
+	require.IsType(t, &privateClient{}, defaultClientFunc("aws_account_id.dkr.ecr.region.amazonaws.com/team-a/app"))
+	require.IsType(t, &publicClient{}, defaultClientFunc("public.ecr.aws/team-a/app"))
+}
+
+func TestECRCredential(t *testing.T) {
+	for _, tt := range []struct {
+		name     string
+		token    string
+		username string
+		password string
+		err      error
+	}{
+		{
+			name:  "invalid base64 token",
+			token: "invalid",
+			err:   base64.CorruptInputError(4),
+		},
+		{
+			name:  "invalid format token",
+			token: "dXNlcl9uYW1lcGFzc3dvcmQ=",
+			err:   auth.ErrBasicCredentialNotFound,
+		},
+		{
+			name:     "valid token",
+			token:    "dXNlcl9uYW1lOnBhc3N3b3Jk",
+			username: "user_name",
+			password: "password",
+		},
+	} {
+		t.Run(tt.name, func(t *testing.T) {
+			r := &CredentialsStore{}
+			credential, err := r.extractCredential(tt.token)
+			assert.Equal(t, tt.err, err)
+			assert.Equal(t, tt.username, credential.Username)
+			assert.Equal(t, tt.password, credential.Password)
+		})
+	}
+}
+
+func TestCredential(t *testing.T) {
+	t.Run("on error", func(t *testing.T) {
+		m := NewMockClient(t)
+		m.On("GetAuthorizationToken", mock.Anything).Return("", time.Time{}, io.ErrUnexpectedEOF)
+		r := &CredentialsStore{
+			cache:      map[string]cacheItem{},
+			clientFunc: func(serverAddress string) Client { return m },
+		}
+		_, err := r.Get(context.Background(), "")
+		assert.ErrorIs(t, err, io.ErrUnexpectedEOF)
+	})
+
+	t.Run("on extract failure", func(t *testing.T) {
+		m := NewMockClient(t)
+		m.On("GetAuthorizationToken", mock.Anything).Return("failure", time.Time{}, nil)
+		r := &CredentialsStore{
+			cache:      map[string]cacheItem{},
+			clientFunc: func(serverAddress string) Client { return m },
+		}
+		_, err := r.Get(context.Background(), "")
+		assert.Error(t, err)
+	})
+
+	t.Run("on success with cached", func(t *testing.T) {
+		m := NewMockClient(t)
+		m.On("GetAuthorizationToken", mock.Anything).Return("dXNlcl9uYW1lOnBhc3N3b3Jk", time.Now().Add(time.Minute), nil).Once()
+		r := &CredentialsStore{
+			cache:      map[string]cacheItem{},
+			clientFunc: func(serverAddress string) Client { return m },
+		}
+		for i := 0; i < 3; i++ {
+			credential, err := r.Get(context.Background(), "")
+			assert.NoError(t, err)
+			assert.Equal(t, "user_name", credential.Username)
+			assert.Equal(t, "password", credential.Password)
+		}
+	})
+}
