+++ internal/oci/ecr/ecr_test.go
@@ -2,9 +2,7 @@ package ecr
 
 import (
 	"context"
-	"encoding/base64"
 	"errors"
-	"io"
 	"testing"
 	"time"
 
@@ -12,76 +10,10 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/ecr/types"
 	"github.com/aws/aws-sdk-go-v2/service/ecrpublic"
 	ptypes "github.com/aws/aws-sdk-go-v2/service/ecrpublic/types"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 	"oras.land/oras-go/v2/registry/remote/auth"
 )
-
-func TestECRCredential(t *testing.T) {
-	for _, tt := range []struct {
-		name     string
-		token    string
-		username string
-		password string
-		err      error
-	}{
-		{
-			name:  "invalid base64 token",
-			token: "invalid",
-			err:   base64.CorruptInputError(4),
-		},
-		{
-			name:  "invalid format token",
-			token: "dXNlcl9uYW1lcGFzc3dvcmQ=",
-			err:   auth.ErrBasicCredentialNotFound,
-		},
-		{
-			name:     "valid token",
-			token:    "dXNlcl9uYW1lOnBhc3N3b3Jk",
-			username: "user_name",
-			password: "password",
-		},
-	} {
-		t.Run(tt.name, func(t *testing.T) {
-			r := &svc{}
-			credential, err := r.extractCredential(tt.token)
-			assert.Equal(t, tt.err, err)
-			assert.Equal(t, tt.username, credential.Username)
-			assert.Equal(t, tt.password, credential.Password)
-		})
-	}
-}
-
-func TestCredential(t *testing.T) {
-	t.Run("on error", func(t *testing.T) {
-		m := NewMockClient(t)
-		m.On("GetAuthorizationToken", mock.Anything).Return("", time.Time{}, io.ErrUnexpectedEOF)
-		r := &svc{client: m}
-		_, err := r.Credential(context.Background(), "")
-		assert.ErrorIs(t, err, io.ErrUnexpectedEOF)
-	})
-
-	t.Run("on extract failure", func(t *testing.T) {
-		m := NewMockClient(t)
-		m.On("GetAuthorizationToken", mock.Anything).Return("failure", time.Time{}, nil)
-		r := &svc{client: m}
-		_, err := r.Credential(context.Background(), "")
-		assert.Error(t, err)
-	})
-
-	t.Run("on success with cached", func(t *testing.T) {
-		m := NewMockClient(t)
-		m.On("GetAuthorizationToken", mock.Anything).Return("dXNlcl9uYW1lOnBhc3N3b3Jk", time.Now().Add(time.Minute), nil).Once()
-		r := &svc{client: m}
-		for i := 0; i < 3; i++ {
-			credential, err := r.Credential(context.Background(), "")
-			assert.NoError(t, err)
-			assert.Equal(t, "user_name", credential.Username)
-			assert.Equal(t, "password", credential.Password)
-		}
-	})
-}
 
 func TestPrivateClient(t *testing.T) {
 	t.Run("on error", func(t *testing.T) {
@@ -170,11 +102,4 @@ func TestPublicClient(t *testing.T) {
 		_, _, err := client.GetAuthorizationToken(context.Background())
 		require.Error(t, err)
 	})
-}
-
-func TestCredentialFunc(t *testing.T) {
-	require.NotNil(t, CredentialFunc("aws_account_id.dkr.ecr.region.amazonaws.com/team-a/app"))
-	require.NotNil(t, CredentialFunc("public.ecr.aws/team-a/app"))
-	require.IsType(t, &privateClient{}, new("aws_account_id.dkr.ecr.region.amazonaws.com/team-a/app").client)
-	require.IsType(t, &publicClient{}, new("public.ecr.aws/team-a/app").client)
 }
